added dpi info from LIB

+++ Pla.Lib/DeviceInfo.cs
@@ -2,6 +2,6 @@ namespace Pla.Lib
 {
     public class DeviceInfo
     {
-        public double DPI{get;set;}
+        public int DPI{get;set;}
     }
 }
